Split NoConstructorArgumentsForInterfaceMock_1
