+++ Source/Moq.Analyzers.Test/NoConstructorArgumentsForInterfaceMockAnalyzerTests.ShouldFailIfMockedInterfaceHasConstructorParameters.verified.txt
@@ -1,0 +1,32 @@
+﻿Diagnostic 1
+	Id: Moq1001
+	Location: SourceFile(Test0.cs[231..241))
+	Highlight: (25, true)
+	Lines: var mock1 = new Mock<IMyService>(25, true);
+	Severity: Warning
+	Message: Mocked interfaces cannot have constructor parameters.
+
+Diagnostic 2
+	Id: Moq1001
+	Location: SourceFile(Test0.cs[283..290))
+	Highlight: ("123")
+	Lines: var mock2 = new Mock<IMyService>("123");
+	Severity: Warning
+	Message: Mocked interfaces cannot have constructor parameters.
+
+Diagnostic 3
+	Id: Moq1001
+	Location: SourceFile(Test0.cs[332..342))
+	Highlight: (25, true)
+	Lines: var mock3 = new Mock<IMyService>(25, true);
+	Severity: Warning
+	Message: Mocked interfaces cannot have constructor parameters.
+
+Diagnostic 4
+	Id: Moq1001
+	Location: SourceFile(Test0.cs[384..391))
+	Highlight: ("123")
+	Lines: var mock4 = new Mock<IMyService>("123");
+	Severity: Warning
+	Message: Mocked interfaces cannot have constructor parameters.
+
