+++ Source/Moq.Analyzers.Test/NoConstructorArgumentsForInterfaceMockAnalyzerTests.ShouldFailIfMockedInterfaceHasConstructorParametersAndExplicitMockBehavior.verified.txt
@@ -1,0 +1,32 @@
+﻿Diagnostic 1
+	Id: Moq1001
+	Location: SourceFile(Test0.cs[263..292))
+	Highlight: (MockBehavior.Default, "123")
+	Lines: var mock1 = new Mock<IMyService>(MockBehavior.Default, "123");
+	Severity: Warning
+	Message: Mocked interfaces cannot have constructor parameters.
+
+Diagnostic 2
+	Id: Moq1001
+	Location: SourceFile(Test0.cs[334..365))
+	Highlight: (MockBehavior.Strict, 25, true)
+	Lines: var mock2 = new Mock<IMyService>(MockBehavior.Strict, 25, true);
+	Severity: Warning
+	Message: Mocked interfaces cannot have constructor parameters.
+
+Diagnostic 3
+	Id: Moq1001
+	Location: SourceFile(Test0.cs[407..436))
+	Highlight: (MockBehavior.Default, "123")
+	Lines: var mock3 = new Mock<IMyService>(MockBehavior.Default, "123");
+	Severity: Warning
+	Message: Mocked interfaces cannot have constructor parameters.
+
+Diagnostic 4
+	Id: Moq1001
+	Location: SourceFile(Test0.cs[478..508))
+	Highlight: (MockBehavior.Loose, 25, true)
+	Lines: var mock4 = new Mock<IMyService>(MockBehavior.Loose, 25, true);
+	Severity: Warning
+	Message: Mocked interfaces cannot have constructor parameters.
+
