--- Source/Moq.Analyzers.Test/NoConstructorArgumentsForInterfaceMockAnalyzerTests.ShouldFailIfMockedInterfaceHasParameters.verified.txt
@@ -1,64 +1,0 @@
-﻿Diagnostic 1
-	Id: Moq1001
-	Location: SourceFile(Test0.cs[225..235))
-	Highlight: (25, true)
-	Lines: var mock1 = new Mock<IMyService>(25, true);
-	Severity: Warning
-	Message: Mocked interfaces cannot have constructor parameters.
-
-Diagnostic 2
-	Id: Moq1001
-	Location: SourceFile(Test0.cs[277..284))
-	Highlight: ("123")
-	Lines: var mock2 = new Mock<IMyService>("123");
-	Severity: Warning
-	Message: Mocked interfaces cannot have constructor parameters.
-
-Diagnostic 3
-	Id: Moq1001
-	Location: SourceFile(Test0.cs[326..336))
-	Highlight: (25, true)
-	Lines: var mock3 = new Mock<IMyService>(25, true);
-	Severity: Warning
-	Message: Mocked interfaces cannot have constructor parameters.
-
-Diagnostic 4
-	Id: Moq1001
-	Location: SourceFile(Test0.cs[378..385))
-	Highlight: ("123")
-	Lines: var mock4 = new Mock<IMyService>("123");
-	Severity: Warning
-	Message: Mocked interfaces cannot have constructor parameters.
-
-Diagnostic 5
-	Id: Moq1001
-	Location: SourceFile(Test0.cs[468..497))
-	Highlight: (MockBehavior.Default, "123")
-	Lines: var mock1 = new Mock<IMyService>(MockBehavior.Default, "123");
-	Severity: Warning
-	Message: Mocked interfaces cannot have constructor parameters.
-
-Diagnostic 6
-	Id: Moq1001
-	Location: SourceFile(Test0.cs[539..570))
-	Highlight: (MockBehavior.Strict, 25, true)
-	Lines: var mock2 = new Mock<IMyService>(MockBehavior.Strict, 25, true);
-	Severity: Warning
-	Message: Mocked interfaces cannot have constructor parameters.
-
-Diagnostic 7
-	Id: Moq1001
-	Location: SourceFile(Test0.cs[612..641))
-	Highlight: (MockBehavior.Default, "123")
-	Lines: var mock3 = new Mock<IMyService>(MockBehavior.Default, "123");
-	Severity: Warning
-	Message: Mocked interfaces cannot have constructor parameters.
-
-Diagnostic 8
-	Id: Moq1001
-	Location: SourceFile(Test0.cs[683..713))
-	Highlight: (MockBehavior.Loose, 25, true)
-	Lines: var mock4 = new Mock<IMyService>(MockBehavior.Loose, 25, true);
-	Severity: Warning
-	Message: Mocked interfaces cannot have constructor parameters.
-
