+++ Source/Moq.Analyzers.Test/NoConstructorArgumentsForInterfaceMockAnalyzerTests.ShouldPassIfMockedInterfaceDoesNotHaveConstructorParameters.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
